backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
hypothesis>=6.100.0
pytest>=8.0.0
openai>=1.0.0
opencv-python>=4.8.0
PyPDF2>=3.0.0
eventlet>=0.35.0
python-engineio>=4.8.0
python-socketio>=5.10.0
